fixed registration test case
