+++ Lib/driver_lib.py
@@ -3,9 +3,9 @@ from selenium import webdriver
 class Driver_Lib:
 
     def get_driver(self):
-        driver = webdriver.Chrome()
-        driver.maximize_window()
-        return driver
+        self.driver = webdriver.Chrome()
+        self.driver.maximize_window()
+        return self.driver
     
     def quite_driver(self):
         self.driver.quit()
