+++ Page/registration.py
@@ -1,7 +1,5 @@
-import logging
 from faker import Faker
 from selenium.webdriver.common.by import By
-from test_data import test_data
 from config import config_data
 from Lib.general_lib import General_Helper
 
@@ -43,13 +41,4 @@ class RegistrationPage(General_Helper):
         return sucess_message
 
 
-    def generate_random_user():
-        fake = Faker()
-        name = fake.name()
-        email = fake.email()
-        username = fake.user_name()
-        password = fake.password()
-        confirm_password = password
-
-        return name, email, username, password, confirm_password
-        return
+
